requirements.txt
pandas==1.5.1
pytest==7.2.0
python_dateutil==2.8.2
